Move session to fixture pkg
